:sparkles:: send 데이터 구조 변화
